gst/va: move VP9 encoder test classes to lib
