+++ test/gst-va/encode/vp9.py
@@ -5,49 +5,10 @@
 ###
 
 from ....lib import *
-from ....lib.codecs import Codec
 from ....lib.gstreamer.va.util import *
-from ....lib.gstreamer.va.encoder import EncoderTest
+from ....lib.gstreamer.va.encoder import VP9EncoderTest, VP9EncoderLPTest
 
 spec = load_test_spec("vp9", "encode", "8bit")
-
-@slash.requires(*have_gst_element("vavp9dec"))
-class VP9EncoderBaseTest(EncoderTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      codec         = Codec.VP9,
-      gstdecoder    = "vavp9dec",
-      gstmediatype  = "video/x-vp9",
-      gstparser     = "vp9parse",
-      gstmuxer      = "matroskamux",
-      gstdemuxer    = "matroskademux",
-    )
-
-  def get_file_ext(self):
-    return "webm"
-
-@slash.requires(*platform.have_caps("encode", "vp9_8"))
-@slash.requires(*have_gst_element("vavp9enc"))
-class VP9EncoderTest(VP9EncoderBaseTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps          = platform.get_caps("encode", "vp9_8"),
-      lowpower      = False,
-      gstencoder    = "vavp9enc",
-    )
-
-@slash.requires(*platform.have_caps("vdenc", "vp9_8"))
-@slash.requires(*have_gst_element("vavp9lpenc"))
-class VP9EncoderLPTest(VP9EncoderBaseTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps          = platform.get_caps("vdenc", "vp9_8"),
-      lowpower      = True,
-      gstencoder    = "vavp9lpenc",
-    )
 
 class cqp_lp(VP9EncoderLPTest):
   def init(self, tspec, case, ipmode, qp, quality, looplvl, loopshp):
